修复System任务休眠后无法唤醒

+++ PMSM_Control_01/Motor_Control/Speed_Loop/Inc/System_Loop.h
@@ -6,9 +6,9 @@
 
 #define SYSTEM_HZ               (1000.0f)
 
-#define SYSTEM_LV_INIT_TIME        (3.0f * SYSTEM_HZ)
-#define SYSTEM_HV_STANDY_TIME   (0.1f * SYSTEM_HZ)
-#define SYSTEM_WAIT_TIME      (3.0f * SYSTEM_HZ)
+#define SYSTEM_LV_INIT_TIME         (uint32_t)(3.0f * SYSTEM_HZ)
+#define SYSTEM_HV_STANDY_TIME       (uint32_t)(0.1f * SYSTEM_HZ)
+#define SYSTEM_WAIT_TIME            (uint32_t)(3.0f * SYSTEM_HZ)
 
 enum SYSTEM_State_t
 {
